requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
Pillow>=10.0
emoji>=2.12
regex>=2024.5.15
datasets>=2.20
pyyaml>=6.0
pyarrow>=16.0
tqdm>=4.66
torch>=2.2
transformers>=4.41
